# regalloc: fast access to vrState