--- a/internal/engine/wazevo/backend/regalloc/regalloc_test.go
+++ b/internal/engine/wazevo/backend/regalloc/regalloc_test.go
@@ -13,10 +13,15 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 	const realRegID, realRegID2 = 50, 100
 	realReg, realReg2 := FromRealReg(realRegID, RegTypeInt), FromRealReg(realRegID2, RegTypeInt)
 	phiVReg := VReg(12345).SetRegType(RegTypeInt)
+
+	type exp struct {
+		liveIns []VRegID
+	}
+
 	for _, tc := range []struct {
 		name  string
 		setup func() Function
-		exp   map[int]*blockState
+		exps  map[int]*exp
 	}{
 		{
 			name: "single block",
@@ -28,7 +33,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 					).entry(),
 				)
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {},
 			},
 		},
@@ -46,7 +51,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 				blk.blockParam(param)
 				return newMockFunction(blk)
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {},
 			},
 		},
@@ -71,7 +76,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 				b1.addPred(b0)
 				return newMockFunction(b0, b1, b2)
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {},
 				1: {
 					liveIns: []VRegID{3},
@@ -112,7 +117,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 				b2.addPred(b0)
 				return newMockFunction(b0, b1, b2, b3)
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {},
 				1: {liveIns: []VRegID{1000, 1}},
 				2: {
@@ -154,7 +159,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 				b1.addPred(b0)
 				return newMockFunction(b0, b1, b2, b3, b4)
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {},
 				1: {
 					liveIns: []VRegID{2000, 3000},
@@ -212,7 +217,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 				f.loopNestingForestRoots(b1)
 				return f
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {
 					liveIns: []VRegID{},
 				},
@@ -253,7 +258,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 				f.loopNestingForestRoots(b2)
 				return f
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {},
 				1: {
 					liveIns: []VRegID{9999},
@@ -303,7 +308,7 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 				f.loopNestingForestRoots(b1)
 				return f
 			},
-			exp: map[int]*blockState{
+			exps: map[int]*exp{
 				0: {
 					liveIns: []VRegID{111},
 				},
@@ -335,17 +340,21 @@ func TestAllocator_livenessAnalysis(t *testing.T) {
 					continue
 				}
 				t.Run(fmt.Sprintf("block_id=%d", blockID), func(t *testing.T) {
-					exp := tc.exp[blockID]
+					exp := tc.exps[blockID]
 					if len(exp.liveIns) == 0 {
 						require.Nil(t, actual.liveIns, "live ins")
 					} else {
-						sort.Slice(actual.liveIns, func(i, j int) bool {
-							return actual.liveIns[i] < actual.liveIns[j]
+						var actuals []VRegID
+						for _, s := range actual.liveIns {
+							actuals = append(actuals, s.v.ID())
+						}
+						sort.Slice(actuals, func(i, j int) bool {
+							return actuals[i] < actuals[j]
 						})
 						sort.Slice(exp.liveIns, func(i, j int) bool {
 							return exp.liveIns[i] < exp.liveIns[j]
 						})
-						require.Equal(t, exp.liveIns, actual.liveIns, "live ins")
+						require.Equal(t, exp.liveIns, actuals, "live ins")
 					}
 				})
 			}
@@ -367,13 +376,13 @@ func TestAllocator_livenessAnalysis_copy(t *testing.T) {
 func Test_findOrSpillAllocatable_prefersSpill(t *testing.T) {
 	t.Run("ok", func(t *testing.T) {
 		s := &state{regsInUse: newRegInUseSet()}
-		s.regsInUse.add(RealReg(1), VReg(2222222))
+		s.regsInUse.add(RealReg(1), &vrState{v: VReg(2222222)})
 		got := s.findOrSpillAllocatable(&Allocator{}, []RealReg{3}, 0, 3)
 		require.Equal(t, RealReg(3), got)
 	})
 	t.Run("preferred but in use", func(t *testing.T) {
 		s := &state{vrStates: wazevoapi.NewIDedPool[vrState](resetVrState)}
-		s.regsInUse.add(RealReg(3), VReg(1).SetRealReg(3))
+		s.regsInUse.add(RealReg(3), &vrState{v: VReg(1).SetRealReg(3)})
 		got := s.findOrSpillAllocatable(&Allocator{}, []RealReg{3, 4}, 0, 3)
 		require.Equal(t, RealReg(4), got)
 	})